refactor: split spells into individual files to prevent merge conflicts
Split each spell into its own file organized by color subdirectories:
- red/ (fireball, flameWall, tremor)
- blue/ (snowstorm, chill)
- green/ (restoration)
- white/ (expose, whirlwind)

This eliminates merge conflicts when multiple agents add spells to the
same color simultaneously. Each new spell is now a new file rather than
an edit to a shared file.

External API unchanged - consumers still import from spells/index.ts.
