+++ packages/core/src/data/spells/red/tremor.ts
@@ -1,89 +1,28 @@
 /**
- * Red spell card definitions
- *
- * Red spells are powered by BLACK + RED mana.
+ * Tremor / Earthquake (Red Spell #11)
+ * Basic: Target enemy gets Armor -3, OR all enemies get Armor -2.
+ * Powered: Target enemy gets Armor -4, OR all enemies get Armor -3.
  */
 
-import type { DeedCard } from "../../types/cards.js";
+import type { DeedCard } from "../../../types/cards.js";
 import {
   CATEGORY_COMBAT,
   DEED_CARD_TYPE_SPELL,
-} from "../../types/cards.js";
+} from "../../../types/cards.js";
 import {
-  EFFECT_COMPOUND,
-  EFFECT_TAKE_WOUND,
   EFFECT_CHOICE,
   EFFECT_SELECT_COMBAT_ENEMY,
   EFFECT_APPLY_MODIFIER,
-} from "../../types/effectTypes.js";
-import {
-  MANA_RED,
-  MANA_BLACK,
-  CARD_FIREBALL,
-  CARD_FLAME_WALL,
-  CARD_TREMOR,
-} from "@mage-knight/shared";
-import type { CardId } from "@mage-knight/shared";
+} from "../../../types/effectTypes.js";
+import { MANA_RED, MANA_BLACK, CARD_TREMOR } from "@mage-knight/shared";
 import {
   DURATION_COMBAT,
   EFFECT_ENEMY_STAT,
   ENEMY_STAT_ARMOR,
   SCOPE_ALL_ENEMIES,
-} from "../../types/modifierConstants.js";
-import {
-  fireRangedAttack,
-  fireSiegeAttack,
-  fireAttack,
-  fireBlock,
-  choice,
-} from "./helpers.js";
+} from "../../../types/modifierConstants.js";
 
-/**
- * Fireball (Red Spell #09)
- * Basic: Ranged Fire Attack 5
- * Powered: Take a Wound. Siege Fire Attack 8.
- */
-const FIREBALL: DeedCard = {
-  id: CARD_FIREBALL,
-  name: "Fireball",
-  poweredName: "Firestorm",
-  cardType: DEED_CARD_TYPE_SPELL,
-  categories: [CATEGORY_COMBAT],
-  poweredBy: [MANA_BLACK, MANA_RED],
-  basicEffect: fireRangedAttack(5),
-  poweredEffect: {
-    type: EFFECT_COMPOUND,
-    effects: [{ type: EFFECT_TAKE_WOUND, amount: 1 }, fireSiegeAttack(8)],
-  },
-  sidewaysValue: 1,
-};
-
-/**
- * Flame Wall / Flame Wave (Red Spell #10)
- * Basic (Flame Wall): Fire Attack 5, or Fire Block 7
- * Powered (Flame Wave): Same choice, +2 per enemy
- *
- * Note: The scaling powered effect is not yet implemented.
- * For now, powered just gives the base values.
- */
-const FLAME_WALL: DeedCard = {
-  id: CARD_FLAME_WALL,
-  name: "Flame Wall",
-  poweredName: "Flame Wave",
-  cardType: DEED_CARD_TYPE_SPELL,
-  categories: [CATEGORY_COMBAT],
-  poweredBy: [MANA_BLACK, MANA_RED],
-  basicEffect: choice([fireAttack(5), fireBlock(7)]),
-  poweredEffect: choice([fireAttack(5), fireBlock(7)]), // TODO: Add scaling
-  sidewaysValue: 1,
-};
-
-/**
- * Tremor / Earthquake (Red Spell #11)
- * Basic: Target enemy gets Armor -3, OR all enemies get Armor -2.
- * Powered: Target enemy gets Armor -4, OR all enemies get Armor -3.
- */
-const TREMOR: DeedCard = {
+export const TREMOR: DeedCard = {
   id: CARD_TREMOR,
   name: "Tremor",
   poweredName: "Earthquake",
@@ -159,10 +98,4 @@ const TREMOR: DeedCard = {
     ],
   },
   sidewaysValue: 1,
-};
-
-export const RED_SPELLS: Record<CardId, DeedCard> = {
-  [CARD_FIREBALL]: FIREBALL,
-  [CARD_FLAME_WALL]: FLAME_WALL,
-  [CARD_TREMOR]: TREMOR,
 };
